% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggallin.R
\docType{package}
\name{ggallin-package}
\alias{ggallin-package}
\title{Grab Bag of GGplot2 Functions.}
\description{
This package consists of some helper functions for working with
\code{ggplot2}: geoms, transforms, \emph{etc.}, with no real
unifying theme among them.
}
\section{Legal Mumbo Jumbo}{


ggallin is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU Lesser General Public License for more details.
}

\author{
Steven E. Pav \email{shabbychef@gmail.com}
}
\keyword{package}
