% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggallin.R
\name{ggallin-NEWS}
\alias{ggallin-NEWS}
\title{News for package 'ggallin':}
\description{
News for package 'ggallin'

\newcommand{\CRANpkg}{\href{https://cran.r-project.org/package=#1}{\pkg{#1}}}
\newcommand{\cranggallin}{\CRANpkg{ggallin}}
\newcommand{\ggallin}{\href{https://github.com/shabbychef/ggallin}}
}
\section{\ggallin{} Version 0.1.2 (2025-11-18) }{

\itemize{
\item fixing figure width styling for CRAN
}
}

\section{\ggallin{} Version 0.1.1 (2017-10-01) }{

\itemize{
\item submit to CRAN
}
}

