% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_subqq.R
\name{stat_subqq}
\alias{stat_subqq}
\title{stat_subqq .}
\usage{
stat_subqq(
  mapping = NULL,
  data = NULL,
  ...,
  position = "dodge",
  na.rm = TRUE,
  distribution = stats::qnorm,
  density = stats::dnorm,
  cuts = 9,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{...}{arguments passed on to ...}

\item{position}{A position adjustment to use on the data for this layer. This
can be used in various ways, including to prevent overplotting and
improving the display. The \code{position} argument accepts the following:
\itemize{
\item The result of calling a position function, such as \code{position_jitter()}.
This method allows for passing extra arguments to the position.
\item A string naming the position adjustment. To give the position as a
string, strip the function name of the \code{position_} prefix. For example,
to use \code{position_jitter()}, give the position as \code{"jitter"}.
\item For more information and other ways to specify the position, see the
\link[ggplot2:layer_positions]{layer position} documentation.
}}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{distribution}{Distribution function to use, if x not specified}

\item{cuts}{vector of blah}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}
}
\description{
A Q-Q plot for large sample sizes.
}
\details{
Implements the large sample Q-Q plot of Velez and Correa Morales,
where the difference between sample quantiles and theoretical
quantiles are plotted versus decile or vigintile number,
along with standard error bars around the \eqn{x} axis.
}
\section{Aesthetics}{

\code{stat_subqq} understands the following aesthetics (required aesthetics
are in bold):
\itemize{
  \item \strong{\code{sample}}
  \item \code{color}
  \item \code{alpha}
}
}

\examples{
\dontrun{
df <- data.frame(y=rnorm(10000))
p <- ggplot(df, aes(sample=y))
p + stat_subqq()
}
}
\references{
Vélez, J. I. and Correa Morales, J. C. "A modified Q-Q plot for large sample
sizes." Communicaciones en Estadística 8, no. 2 (2015): 163-172. 
\url{https://revistas.usantotomas.edu.co/index.php/estadistica/article/download/1441/2453}
}
\author{
Steven E. Pav \email{shabbychef@gmail.com}
}
\keyword{distribution}
\keyword{plotting}
