% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\docType{data}
\name{ssqrt_trans}
\alias{ssqrt_trans}
\alias{pseudolog10_trans}
\alias{interp_trans}
\alias{warp_trans}
\title{transforms .}
\format{An object of class \code{trans} of length 6.}
\usage{
ssqrt_trans

pseudolog10_trans

interp_trans(x, y, na.rm = TRUE)

warp_trans(x, w, na.rm = TRUE)
}
\arguments{
\item{...}{arguments passed on to ...
Invalid arguments will result in return value \code{NaN} with a warning.}
}
\value{
foo
}
\description{
Various scale transformations.
}
\details{
The available transforms:
\itemize{
\item \code{ssqrt_trans} a signed square root transform appropriate for
negative or positive numbers.
\item \code{pseudolog10_trans} an \code{asinh} transformation, which is like
a logarithm, but appropriate for negative or positive numbers.
}
}
\examples{
ggplot(data.frame(x=rnorm(100),y=runif(100)),aes(x=x,y=y)) + geom_point() + scale_x_continuous(trans=ssqrt_trans)

ggplot(data.frame(x=rnorm(100),y=runif(100)),aes(x=x,y=y)) + 
  geom_point() + 
  scale_x_continuous(trans=interp_trans(x=seq(-10,10,by=1),y=cumsum(runif(21))))

ggplot(data.frame(x=rnorm(100),y=runif(100)),aes(x=x,y=y)) + 
  geom_point() + 
  scale_x_continuous(trans=warp_trans(x=seq(-10,10,by=1),y=runif(21)))
}
\seealso{
...

\url{http://www.win-vector.com/blog/2012/03/modeling-trick-the-signed-pseudo-logarithm/}
}
\author{
Steven E. Pav \email{steven@gilgamath.com}
}
\keyword{datasets}
\keyword{plotting}
